Move temporary message functionality to another file
